+++ apps/ai-dial-admin/src/components/Applications/View/Properties/Properties.tsx
@@ -1,12 +1,12 @@
 import { FC } from 'react';
 
-import Defaults from '@/src/components/Defaults/Defaults';
 import IconControl from '@/src/components/BaseControls/Icon';
 import MaxRetryAttempts from '@/src/components/BaseControls/MaxRetryAttempts';
 import TopicsControl from '@/src/components/BaseControls/Topics';
+import Defaults from '@/src/components/Defaults/Defaults';
 import EntityAttachments from '@/src/components/EntityMainProperties/EntityAttachments/EntityAttachments';
 import ForwardAuthTokenField from '@/src/components/EntityMainProperties/ForwardAuthToken/ForwardAuthTokenField';
-import Properties from '@/src/components/EntityMainProperties/Properties/Properties';
+import DeploymentProperties from '@/src/components/EntityMainProperties/Properties/DeploymentProperties';
 import ApplicationSource from '@/src/components/SourceField/Application/ApplicationSource';
 import { DialApplicationScheme } from '@/src/models/dial/application';
 import { ChatEntity } from '@/src/models/dial/base-entity';
@@ -20,29 +20,20 @@ interface Props {
   onChangeEntity: (entity: ChatEntity) => void;
 }
 
-const EntityProperties: FC<Props> = ({ runners, names, view, ...props }) => {
+const EntityProperties: FC<Props> = ({ runners, view, ...props }) => {
   return (
     <div className="h-full flex flex-col gap-y-8">
-      <Properties {...props} view={view} names={names} runners={runners} isEntityImmutable={true} />
-
+      <DeploymentProperties view={view} runners={runners} isUniqueNameError={false} {...props} />
       <IconControl
         iconUrl={props.entity.iconUrl}
         onChange={(icon) => props.onChangeEntity({ ...props.entity, iconUrl: icon })}
       />
-
       <TopicsControl {...props} onChange={props.onChangeEntity} />
-
-      {view === ApplicationRoute.Applications && (
-        <ApplicationSource {...props} runners={runners} isEntityImmutable={true} />
-      )}
-
+      <ApplicationSource {...props} runners={runners} isEntityImmutable={true} />
       <EntityAttachments {...props} />
-
-      {view === ApplicationRoute.Applications && <Defaults {...props} />}
-
+      && <Defaults {...props} />
       <ForwardAuthTokenField view={view} {...props} />
-
-      {view === ApplicationRoute.Applications && <MaxRetryAttempts {...props} />}
+      <MaxRetryAttempts {...props} />
     </div>
   );
 };
