+++ apps/ai-dial-admin/src/components/Models/View/TabsContent.tsx
@@ -3,16 +3,15 @@ import { FC } from 'react';
 
 import EntityAudit from '@/src/components/EntityTabs/Audit/EntityAudit';
 import EntityFeatures from '@/src/components/EntityTabs/Features/Features';
+import PropertiesTabContent from '@/src/components/EntityTabs/PropertiesTabContent';
 import EntityInterceptors from '@/src/components/EntityView/Interceptors/Interceptors';
 import EntityRoles from '@/src/components/EntityView/Roles/Roles';
-import PropertiesContent from '@/src/components/EntityView/View/Content/PropertiesContent';
+import ModelProperties from '@/src/components/ModelView/ModelProperties/ModelProperties';
 import { DialInterceptor } from '@/src/models/dial/interceptor';
 import { DialModel } from '@/src/models/dial/model';
 import { DialRole } from '@/src/models/dial/role';
 import { ApplicationRoute } from '@/src/types/routes';
 import { EntityViewTab } from '@/src/utils/tabs/utils';
-import PropertiesTabContent from '../../EntityTabs/PropertiesTabContent';
-import ModelProperties from '../../ModelView/ModelProperties/ModelProperties';
 
 interface Props {
   activeTab: EntityViewTab;
